+++ src/containers/StripePayoutPage/StripePayoutPage.js
@@ -3,23 +3,18 @@ import { compose } from 'redux';
 import { connect } from 'react-redux';
 
 import { useConfiguration } from '../../context/configurationContext';
-import { useRouteConfiguration } from '../../context/routeConfigurationContext';
 import { createResourceLocatorString } from '../../util/routes';
 import { FormattedMessage, useIntl } from '../../util/reactIntl';
 import { ensureCurrentUser } from '../../util/data';
-import { propTypes } from '../../util/types';
 import { showCreateListingLinkForUser, showPaymentDetailsForUser } from '../../util/userHelpers';
 import { isScrollingDisabled } from '../../ducks/ui.duck';
-import {
-  stripeAccountClearError,
-  getStripeConnectAccountLink,
-} from '../../ducks/stripeConnectAccount.duck';
+
+import ManualPayoutForm from './ManualPayoutForm';
 
 import {
   H3,
   NamedRedirect,
   Page,
-  StripeConnectAccountStatusBox,
   StripeConnectAccountForm,
   UserNav,
   LayoutSideNavigation,
@@ -81,14 +76,9 @@ const handleGetStripeConnectAccountLinkFn = (getLinkFn, commonParams) => type =>
  *
  * @component
  * @param {Object} props
- * @param {propTypes.currentUser} props.currentUser - The current user
  * @param {boolean} props.scrollingDisabled - Whether scrolling is disabled
  * @param {boolean} props.getAccountLinkInProgress - Whether the account link is in progress
  * @param {boolean} props.payoutDetailsSaveInProgress - Whether the payout details are in progress
- * @param {propTypes.error} props.createStripeAccountError - The create stripe account error
- * @param {propTypes.error} props.getAccountLinkError - The get account link error
- * @param {propTypes.error} props.updateStripeAccountError - The update stripe account error
- * @param {propTypes.error} props.fetchStripeAccountError - The fetch stripe account error
  * @param {Object} props.stripeAccount - The stripe account
  * @param {boolean} props.stripeAccountFetched - Whether the stripe account is fetched
  * @param {boolean} props.payoutDetailsSaved - Whether the payout details are saved
@@ -99,165 +89,46 @@ const handleGetStripeConnectAccountLinkFn = (getLinkFn, commonParams) => type =>
  * @param {STRIPE_ONBOARDING_RETURN_URL_SUCCESS | STRIPE_ONBOARDING_RETURN_URL_FAILURE} props.params.returnURLType - The return URL type (success or failure)
  * @returns {JSX.Element}
  */
-export const StripePayoutPageComponent = props => {
-  const config = useConfiguration();
-  const routes = useRouteConfiguration();
-  const intl = useIntl();
-  const {
-    currentUser,
-    scrollingDisabled,
-    getAccountLinkInProgress,
-    getAccountLinkError,
-    createStripeAccountError,
-    updateStripeAccountError,
-    fetchStripeAccountError,
-    stripeAccountFetched,
-    stripeAccount,
-    onPayoutDetailsChange,
-    onPayoutDetailsSubmit,
-    onGetStripeConnectAccountLink,
-    payoutDetailsSaveInProgress,
-    payoutDetailsSaved,
-    params,
-    authScopes,
-  } = props;
 
-  const { returnURLType } = params || {};
+export const StripePayoutPageComponent = props => {
+  const { currentUser, scrollingDisabled, payoutDetailsSaveInProgress, payoutDetailsSaved, onPayoutDetailsSubmit, onPayoutDetailsDelete } = props;  
+  const intl = useIntl();
+  const config = useConfiguration();
+
   const ensuredCurrentUser = ensureCurrentUser(currentUser);
   const currentUserLoaded = !!ensuredCurrentUser.id;
-  const stripeConnected = currentUserLoaded && !!stripeAccount && !!stripeAccount.id;
-
-  const title = intl.formatMessage({ id: 'StripePayoutPage.title' });
-
-  const formDisabled = getAccountLinkInProgress;
-
-  const rootURL = config.marketplaceRootURL;
-  const successURL = createReturnURL(STRIPE_ONBOARDING_RETURN_URL_SUCCESS, rootURL, routes);
-  const failureURL = createReturnURL(STRIPE_ONBOARDING_RETURN_URL_FAILURE, rootURL, routes);
-
-  const accountId = stripeConnected ? stripeAccount.id : null;
-  const stripeAccountData = stripeConnected ? getStripeAccountData(stripeAccount) : null;
-  const requirementsMissing =
-    stripeAccount &&
-    (hasRequirements(stripeAccountData, 'past_due') ||
-      hasRequirements(stripeAccountData, 'currently_due'));
-
-  const savedCountry = stripeAccountData ? stripeAccountData.country : null;
-  const savedAccountType = stripeAccountData ? stripeAccountData.business_type : null;
-
-  const handleGetStripeConnectAccountLink = handleGetStripeConnectAccountLinkFn(
-    onGetStripeConnectAccountLink,
-    {
-      accountId,
-      successURL,
-      failureURL,
-    }
-  );
-
-  const returnedNormallyFromStripe = returnURLType === STRIPE_ONBOARDING_RETURN_URL_SUCCESS;
-  const returnedAbnormallyFromStripe = returnURLType === STRIPE_ONBOARDING_RETURN_URL_FAILURE;
-  const showVerificationNeeded = stripeConnected && requirementsMissing;
-
-  // Check if user has limited rights and set button titles accordingly
-  const limitedRights = authScopes?.indexOf('user:limited') >= 0;
-  const stripeButtonTitle = limitedRights
-    ? intl.formatMessage({ id: 'StripePayoutPage.submitButtonText' })
-    : null;
-
-  // Redirect from success URL to basic path for StripePayoutPage
-  if (returnedNormallyFromStripe && stripeConnected && !requirementsMissing) {
-    return <NamedRedirect name="StripePayoutPage" />;
-  }
-
-  // Failure url should redirect back to Stripe since it's most likely due to page reload
-  // Account link creation will fail if the account is the reason
-  if (returnedAbnormallyFromStripe && !getAccountLinkError) {
-    handleGetStripeConnectAccountLink('custom_account_verification')();
-  }
+  const saved = ensuredCurrentUser.attributes?.profile?.privateData?.manualPayoutDetails;
 
   const showManageListingsLink = showCreateListingLinkForUser(config, currentUser);
   const { showPayoutDetails, showPaymentMethods } = showPaymentDetailsForUser(config, currentUser);
-  const accountSettingsNavProps = {
-    currentPage: 'StripePayoutPage',
-    showPaymentMethods,
-    showPayoutDetails,
-  };
 
   return (
-    <Page title={title} scrollingDisabled={scrollingDisabled}>
+    <Page title="Payout details" scrollingDisabled={scrollingDisabled}>
       <LayoutSideNavigation
         topbar={
           <>
-            <TopbarContainer
-              desktopClassName={css.desktopTopbar}
-              mobileClassName={css.mobileTopbar}
-            />
-            <UserNav
-              currentPage="StripePayoutPage"
-              showManageListingsLink={showManageListingsLink}
-            />
+            <TopbarContainer desktopClassName={css.desktopTopbar} mobileClassName={css.mobileTopbar} />
+            <UserNav currentPage="StripePayoutPage" showManageListingsLink={showManageListingsLink} />
           </>
         }
         sideNav={null}
         useAccountSettingsNav
-        accountSettingsNavProps={accountSettingsNavProps}
+        accountSettingsNavProps={{ currentPage: 'StripePayoutPage', showPaymentMethods, showPayoutDetails }}
         footer={<FooterContainer />}
         intl={intl}
       >
         <div className={css.content}>
-          <H3 as="h1" className={css.heading}>
-            <FormattedMessage id="StripePayoutPage.heading" />
-          </H3>
+          <H3 as="h1" className={css.heading}>Payout details</H3>
           {!currentUserLoaded ? (
-            <FormattedMessage id="StripePayoutPage.loadingData" />
-          ) : returnedAbnormallyFromStripe && !getAccountLinkError ? (
-            <FormattedMessage id="StripePayoutPage.redirectingToStripe" />
+            <p>Loading…</p>
           ) : (
-            <StripeConnectAccountForm
-              rootClassName={css.stripeConnectAccountForm}
-              disabled={formDisabled}
-              inProgress={payoutDetailsSaveInProgress}
-              ready={payoutDetailsSaved}
-              currentUser={ensuredCurrentUser}
-              stripeBankAccountLastDigits={getBankAccountLast4Digits(stripeAccountData)}
-              savedCountry={savedCountry}
-              savedAccountType={savedAccountType}
-              submitButtonText={intl.formatMessage({
-                id: 'StripePayoutPage.submitButtonText',
-              })}
-              stripeAccountError={
-                createStripeAccountError || updateStripeAccountError || fetchStripeAccountError
-              }
-              stripeAccountLinkError={getAccountLinkError}
-              stripeAccountFetched={stripeAccountFetched}
-              onChange={onPayoutDetailsChange}
+            <ManualPayoutForm
               onSubmit={onPayoutDetailsSubmit}
-              onGetStripeConnectAccountLink={handleGetStripeConnectAccountLink}
-              stripeConnected={stripeConnected}
-              authScopes={authScopes}
-            >
-              {stripeConnected && !returnedAbnormallyFromStripe && showVerificationNeeded ? (
-                <StripeConnectAccountStatusBox
-                  type="verificationNeeded"
-                  inProgress={getAccountLinkInProgress}
-                  onGetStripeConnectAccountLink={handleGetStripeConnectAccountLink(
-                    'custom_account_verification'
-                  )}
-                  disabled={limitedRights}
-                  title={stripeButtonTitle}
-                />
-              ) : stripeConnected && savedCountry && !returnedAbnormallyFromStripe ? (
-                <StripeConnectAccountStatusBox
-                  type="verificationSuccess"
-                  inProgress={getAccountLinkInProgress}
-                  disabled={payoutDetailsSaveInProgress || limitedRights}
-                  onGetStripeConnectAccountLink={handleGetStripeConnectAccountLink(
-                    'custom_account_update'
-                  )}
-                  title={stripeButtonTitle}
-                />
-              ) : null}
-            </StripeConnectAccountForm>
+              onDelete={onPayoutDetailsDelete}
+              inProgress={payoutDetailsSaveInProgress}
+              saved={payoutDetailsSaved}
+              existingDetails={saved}
+            />
           )}
         </div>
       </LayoutSideNavigation>
@@ -266,39 +137,19 @@ export const StripePayoutPageComponent = props => {
 };
 
 const mapStateToProps = state => {
-  const {
-    getAccountLinkInProgress,
-    getAccountLinkError,
-    createStripeAccountError,
-    updateStripeAccountError,
-    fetchStripeAccountError,
-    stripeAccount,
-    stripeAccountFetched,
-  } = state.stripeConnectAccount;
   const { currentUser } = state.user;
   const { payoutDetailsSaveInProgress, payoutDetailsSaved } = state.StripePayoutPage;
-  const { authScopes } = state.auth;
   return {
     currentUser,
-    getAccountLinkInProgress,
-    getAccountLinkError,
-    createStripeAccountError,
-    updateStripeAccountError,
-    fetchStripeAccountError,
-    stripeAccount,
-    stripeAccountFetched,
     payoutDetailsSaveInProgress,
     payoutDetailsSaved,
     scrollingDisabled: isScrollingDisabled(state),
-    authScopes,
   };
 };
 
 const mapDispatchToProps = dispatch => ({
-  onPayoutDetailsChange: () => dispatch(stripeAccountClearError()),
-  onPayoutDetailsSubmit: (values, isUpdateCall) =>
-    dispatch(savePayoutDetails(values, isUpdateCall)),
-  onGetStripeConnectAccountLink: params => dispatch(getStripeConnectAccountLink(params)),
+  onPayoutDetailsSubmit: values => dispatch(savePayoutDetails(values)),
+  onPayoutDetailsDelete: () => dispatch(savePayoutDetails(null)),
 });
 
 const StripePayoutPage = compose(