+++ src/containers/TopbarContainer/Topbar/TopbarSearchForm/IconSearchDesktop.js
@@ -5,23 +5,20 @@ import css from './TopbarSearchForm.module.css';
 const IconSearchDesktop = () => (
   <svg
     className={css.iconSvg}
-    width="21"
-    height="22"
-    viewBox="0 0 21 22"
+    width="15"
+    height="15"
+    viewBox="0 0 24 24"
     xmlns="http://www.w3.org/2000/svg"
     role="none"
   >
     <g
       className={css.iconSvgGroup}
-      transform="matrix(-1 0 0 1 20 1)"
-      strokeWidth="2"
       fill="none"
-      fillRule="evenodd"
       strokeLinecap="round"
       strokeLinejoin="round"
     >
-      <path d="M13 14l5.241 5.241" />
-      <circle cx="7.5" cy="7.5" r="7.5" />
+      <circle cx="10" cy="10" r="7" />
+      <line x1="15.5" y1="15.5" x2="21" y2="21" />
     </g>
   </svg>
 );